wasm ducktape: use Cli tools in producer for wasm tests
